+++ scripts/us_cdc/environmental_health_toxicology/parse_precipitation_index_test.py
@@ -14,36 +14,36 @@
 '''
 Author: Padma Gundapaneni @padma-g
 Date: 7/28/21
-Description: This script contains unit tests for the parse_air_quality.py script.
+Description: This script contains unit tests for the parse_precipitation_index.py script.
 @input_file   filepath to the original csv that needs to be cleaned
 @output_file  filepath to the csv to which the cleaned data is written
-python3 test_parse_air_quality.py input_file output_file
+python3 parse_precipitation_index_test.py input_file output_file
 '''
 
 import unittest
 import os
-from parse_air_quality import clean_air_quality_data
+from parse_precipitation_index import clean_precipitation_data
 
 module_dir_ = os.path.dirname(__file__)
 
 
-class TestParseAirQuality(unittest.TestCase):
+class TestParsePrecipitationData(unittest.TestCase):
     """
-    Tests the functions in parse_air_quality.py.
+    Tests the functions in parse_precipitation_index.py.
     """
 
-    def test_clean_air_quality_data(self):
+    def test_clean_precipitation_data(self):
         """
-        Tests the clean_air_quality_data function.
+        Tests the clean_precipitation_data function.
         """
         print(module_dir_)
-        test_csv = os.path.join(module_dir_, 'test_data/small_Ozone_County.csv')
+        test_csv = os.path.join(module_dir_, 'test_data/small_Palmer.csv')
         output_csv = os.path.join(module_dir_,
-                                  'test_data/small_Ozone_County_output.csv')
-        clean_air_quality_data(test_csv, output_csv)
+                                  'test_data/small_Palmer_output.csv')
+        clean_precipitation_data(test_csv, output_csv)
 
-        expected_csv = os.path.join(
-            module_dir_, 'test_data/small_Ozone_County_expected.csv')
+        expected_csv = os.path.join(module_dir_,
+                                    'test_data/small_Palmer_expected.csv')
         with open(output_csv, 'r') as test:
             test_str: str = test.read()
             with open(expected_csv, 'r') as expected:
